Switch catchall middleware to POST route at /waitlist
Ref #23
