+++ src/index.html
@@ -39,7 +39,7 @@
             <div
               class="p-4 p-md-5 border rounded-3 shadow waitlist-form-container"
             >
-              <form data-static-form-name="waitlist" id="waitlist-form">
+              <form id="waitlist-form" action="/waitlist">
                 <input
                   type="text"
                   name="address"
